requirements.txt
amqp==5.2.0
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
beautifulsoup4==4.12.3
billiard==4.2.0
bs4==0.0.2
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
cheap-repr==0.5.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0.2
djangorestframework==3.14.0
executing==2.0.1
flower==2.0.1
gunicorn==21.2.0
humanize==4.9.0
idna==3.6
kombu==5.3.5
packaging==23.2
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psutil==5.9.8
Pygments==2.17.2
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.2
requests==2.31.0
six==1.16.0
snoop==0.4.3
soupsieve==2.5
sqlparse==0.4.4
tornado==6.4
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
websockets==12.0
whitenoise==6.6.0
